requirement.txt
absl-py==0.4.1
alabaster==0.7.10
altgraph==0.15
anaconda-client==1.6.14
anaconda-navigator==1.9.12
anaconda-project==0.8.2
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.1
astropy==2.0.3
attrs==17.4.0
audioread==2.1.8
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
bleach==1.5.0
bokeh==0.12.13
boto==2.48.0
boto3==1.9.168
botocore==1.12.168
Bottleneck==1.2.1
bs4==0.0.1
bz2file==0.98
certifi==2019.11.28
cffi==1.11.4
chardet==3.0.4
chromaprint==0.5
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.4
conda==4.4.10
conda-build==3.4.1
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.1.4
cx-Oracle==7.0.0
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0
dask==0.16.1
datashape==0.5.4
decorator==4.2.1
distributed==1.20.2
Django==2.1.4
docutils==0.14
easydict==1.7
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
ffmpeg==1.4
filelock==2.0.13
Flask==1.0.2
Flask-Cors==3.0.3
future==0.16.0
gast==0.2.0
gensim==3.4.0
gevent==1.2.2
glob2==0.6
greenlet==0.4.12
grpcio==1.12.1
h5py==2.7.1
heapdict==1.0.0
html5lib==0.9999999
idna==2.6
imageio==2.2.0
imageio-ffmpeg==0.3.0
imagesize==0.7.1
ipykernel==5.1.4
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itchat==1.3.10
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
jieba==0.39
Jinja2==2.10
jmespath==0.9.4
jqdatasdk==1.6.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.31.4
jupyterlab-launcher==0.10.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lazy-object-proxy==1.3.1
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
macholib==1.9
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
mccabe==0.6.1
menuinst==1.4.11
mistune==0.8.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==2.0.0
moviepy==1.0.0
mpmath==1.0.0
msgpack==0.6.1
msgpack-python==0.5.1
multipledispatch==0.4.9
mysql-connector-python==8.0.18
navigator-updater==0.1.0
nb-conda==2.2.1
nb-conda-kernels==2.2.2
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.4.0
numba==0.36.2
numexpr==2.6.4
numpy==1.18.1
numpydoc==0.7.0
odo==0.5.1
olefile==0.45.1
opencv-contrib-python==3.3.0.10
opencv-python==3.4.0.12
openpyxl==2.4.10
packaging==16.8
pandas==0.24.2
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pbr==5.1.3
pefile==2017.11.5
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.0.0
pkginfo==1.4.1
pluggy==0.6.0
ply==3.10
proglog==0.1.9
prompt-toolkit==1.0.15
protobuf==3.6.0
psutil==5.4.3
py==1.5.2
pyacoustid==1.1.7
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pydub==0.23.1
pyflakes==1.6.0
pygame==1.9.3
Pygments==2.2.0
PyInstaller==3.3.1
pylint==1.8.2
PyMySQL==0.9.3
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
pypng==0.0.18
PyQRCode==1.2.1
pyrouge==0.1.3
PySocks==1.6.7
pytesseract==0.2.6
pytest==3.3.2
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==1.1.1
pywin32==222
pywinpty==0.5
PyYAML==5.3
pyzmq==16.0.3
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.5.0
requests==2.18.4
rope==0.10.7
ruamel-yaml==0.15.35
s3transfer==0.2.1
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.3.2
seaborn==0.8.1
Send2Trash==1.4.2
seq2seq==0.1
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.8.4
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.2.6
SQLAlchemy==1.3.4
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thriftpy2==0.4.3
toolz==0.9.0
tornado==4.5.3
tqdm==4.31.1
traitlets==4.3.2
tushare==1.2.51
typing==3.6.2
unicodecsv==0.14.1
UNKNOWN==0.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wordcloud==1.5.0
wordninja==2.0.0
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.5
xlwt==1.3.0
xmltodict==0.12.0
zict==0.1.3
